Backgammon-Bot.java
An AI to play Backgammon against other players and bots. Written by Joanne Reilly (joreillyIE) to be run along side Backgammon App. The bot generates all possible plays and rates each possible play based on the probability that it will lead to a win, based on several
weighted scores for each feature. It then picks the play with the highest probability.